refactor: move validators to common file
