+++ eox_nelp/user_profile/required_fields_validation.py
@@ -63,17 +63,54 @@ REQUIRED_USER_FIELDS = {
 }
 """
 import logging
-import re
 
 from django.conf import settings
-from phonenumber_field.phonenumber import PhoneNumber
+
+from eox_nelp.validators import validate_char_type, validate_format, validate_max_length, validate_optional_values
 
 logger = logging.getLogger(__name__)
 
 
 def validate_required_user_fields(user):
     """
-    Validates user fields and returns a dictionary with invalid fields categorized by account, profile, and extra_info.
+    Validates the required fields of a user instance based on the REQUIRED_USER_FIELDS setting.
+
+    This function checks the user's account, profile, and extra_info fields to ensure they comply with the validation
+    rules defined in the settings. If a field is missing or contains invalid data, it is added to the output dictionary
+    under its respective category.
+
+    Args:
+        user (User): The user instance to validate.
+
+    Returns:
+        dict: A dictionary containing invalid fields categorized by "account", "profile", and "extra_info".
+              Each field will have an empty list if valid, or a list of error messages if invalid.
+
+    Example:
+        Given the following REQUIRED_USER_FIELDS setting:
+
+        REQUIRED_USER_FIELDS = {
+            "account": {"first_name": {"max_length": 30}, "last_name": {"max_length": 50}},
+            "profile": {
+                "city": {"max_length": 32},
+                "country": {"max_length": 2},
+                "phone_number": {"format": "phone"},
+                "mailing_address": {"max_length": 40},
+            },
+            "extra_info": {
+                "arabic_name": {"char_type": "arabic"},
+                "arabic_first_name": {"char_type": "arabic"},
+                "arabic_last_name": {"char_type": "arabic"},
+            },
+        }
+
+        A possible output from this function could be:
+
+        {
+            "account": {"first_name": ["An error"], "last_name": []},
+            "profile": {"city": [], "country": [], "phone_number": ["Empty field"], "mailing_address": []},
+            "extra_info": {"arabic_name": [], "arabic_first_name": [], "arabic_last_name": []}
+        }
     """
     required_fields_config = getattr(settings, "REQUIRED_USER_FIELDS", {})
 
@@ -190,165 +227,3 @@ def validate_field(value, rules):
             errors.append(f"{rule} with argument {argument} failed")
 
     return errors
-
-
-def validate_max_length(value, max_length):
-    """
-    Checks if the given value does not exceed the specified max_length.
-
-    Args:
-        value (str): The value to be validated.
-        max_length (int): The maximum allowed length.
-
-    Returns:
-        bool: True if valid, False if it exceeds max_length.
-    """
-    return len(str(value)) <= max_length
-
-
-def validate_optional_values(value, allowed_values):
-    """
-    Checks if the given value is within the predefined allowed values.
-
-    Args:
-        value (str): The value to be validated.
-        allowed_values (list): A list of allowed values.
-
-    Returns:
-        bool: True if the value is in the allowed list, False otherwise.
-    """
-    return value in allowed_values
-
-
-def validate_char_type(value, char_type):
-    """
-    Validates if the given value matches the specified character type.
-
-    Args:
-        value (str): The value to be validated.
-        char_type (str): The character type to validate against. Options: "latin", "arabic", "cp1252".
-
-    Returns:
-        bool: True if the value matches the specified character type, False otherwise.
-    """
-    validators = {
-        "latin": validate_latin,
-        "arabic": validate_arabic,
-        "cp1252": validate_cp1252,
-    }
-
-    validator = validators.get(char_type)
-
-    if validator:
-        return validator(value)
-
-    return False
-
-
-def validate_latin(value):
-    """
-    Checks if the given value contains only Latin characters and standard punctuation.
-
-    Args:
-        value (str): The value to be validated.
-
-    Returns:
-        bool: True if valid, False otherwise.
-    """
-    latin_regex = r'^[a-zA-ZÀ-ÿ\s.,\'"-]*$'
-    return re.match(latin_regex, value) is not None
-
-
-def validate_arabic(value):
-    """
-    Checks if the given value contains only Arabic script characters.
-
-    Args:
-        value (str): The value to be validated.
-
-    Returns:
-        bool: True if valid, False otherwise.
-    """
-    arabic_regex = r'^[\u0600-\u06FF\s]+$'  # Unicode range for Arabic characters
-    return re.match(arabic_regex, value) is not None
-
-
-def validate_cp1252(value):
-    """
-    Checks if the given value contains only Windows-1252 compatible characters.
-
-    Args:
-        value (str): The value to be validated.
-
-    Returns:
-        bool: True if valid, False otherwise.
-    """
-    cp1252_regex = r'^[\x00-\x7F\x80-\x9F\xA0-\xFF]*$'
-    return re.match(cp1252_regex, value) is not None
-
-
-def validate_format(value, field_format):
-    """
-    Validates if the given value matches the specified format.
-
-    Args:
-        value (str): The value to be validated.
-        field_format (str): The format to validate against. Options: "email", "phone", "numeric", "alphanumeric".
-
-    Returns:
-        bool: True if the value matches the specified format, False otherwise.
-    """
-    validators = {
-        "email": validate_email,
-        "phone": validate_phone,
-        "numeric": validate_numeric,
-    }
-
-    validator = validators.get(field_format)
-
-    if validator:
-        return validator(value)
-
-    return False
-
-
-def validate_email(value):
-    """
-    Checks if the given value is a valid email address.
-
-    Args:
-        value (str): The value to be validated.
-
-    Returns:
-        bool: True if valid, False otherwise.
-    """
-    email_regex = r'^[\w\.-]+@[\w\.-]+\.\w+$'
-    return re.match(email_regex, value) is not None
-
-
-def validate_phone(value):
-    """
-    Checks if the given value is a valid phone number using django-phonenumber-field.
-
-    Args:
-        value (str): The phone number to be validated.
-
-    Returns:
-        bool: True if valid, False otherwise.
-    """
-    phone = PhoneNumber.from_string(value)
-
-    return phone.is_valid()
-
-
-def validate_numeric(value):
-    """
-    Checks if the given value contains only numeric digits.
-
-    Args:
-        value (str): The value to be validated.
-
-    Returns:
-        bool: True if valid, False otherwise.
-    """
-    return value.isdigit()
